src/requirements.txt
black==24.4.2
flask==3.0.3
flake8==7.1.0
isort==5.13.2
requests==2.31.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
